move initial files to folders
