+++ .gitignore
@@ -1,1 +1,2 @@
 *.o
+.ccls-cache
