+++ src/drw/drw.c
@@ -6,7 +6,7 @@
 #include <X11/Xft/Xft.h>
 
 #include "drw.h"
-#include "util.h"
+#include "../util.h"
 
 #define UTF_INVALID 0xFFFD
 #define UTF_SIZ     4
