+++ src/dwm.c
@@ -20,32 +20,9 @@
  *
  * To understand everything else, start reading main().
  */
-#include <errno.h>
-#include <locale.h>
-#include <signal.h>
-#include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <X11/cursorfont.h>
-#include <X11/keysym.h>
-#include <X11/Xatom.h>
-#include <X11/Xlib.h>
-#include <X11/Xproto.h>
-#include <X11/Xutil.h>
 
-#ifdef XINERAMA
-
-#include <X11/extensions/Xinerama.h>
-
-#endif /* XINERAMA */
-
-#include <X11/Xft/Xft.h>
-
-#include "drw.h"
+#include "dwm.h"
+#include "drw/drw.h"
 #include "util.h"
 
 /* macros */
@@ -381,7 +358,7 @@ static Monitor *mons, *selmon;
 static Window root, wmcheckwin;
 
 /* configuration, allows nested code to access above variables */
-#include "config.h"
+#include "../config.h"
 
 /* compile-time check if all tags fit into an unsigned int bit array. */
 struct NumTags {
